requirements.txt
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.108.0
filelock==3.13.1
fsspec==2023.12.2
h11==0.14.0
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
omegaconf==2.3.0
pillow==10.2.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pygame==2.5.2
PyYAML==6.0.1
requests==2.31.0
silero @ git+https://github.com/snakers4/silero-models.git@6b0bb8a7637d791fbb7adf22c56af1c89758ff19
sniffio==1.3.0
soundfile==0.12.1
starlette==0.32.0.post1
sympy==1.12
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
